requirements.txt
crewai==1.5.0
google-generativeai==0.11.0
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.12.0
python-dotenv==1.0.0
pydantic-core==2.12.0
instagrapi==2.1.2
Pillow==10.1.0
